created feature_search.py script
