+++ library_filter.py
@@ -1,30 +1,13 @@
+#!/usr/bin/python
+'''
+Title: Library filter functions (main.py)
+Author: Jasen M. Jackson, Loyola '19
+This script contains functions necessary for filtering TRAPseq (teaseq???) libraries.
+'''
 import glob, os
 import os.path
 import sys
-
-#Return a list of Kmers that cover "threshold" % of the primerss
-def kmers_threshold(primer, threshold):
-
-	k = int(round(threshold*len(primer))) #k_min = 0.9*primer.length
-	kmers = []
-	for i in range(len(primer)-k+1):
-		kmer = primer[i:i+k]
-		kmers.append(kmer)
-	return(kmers, k)
-
-#Returns a list of Kmers of length k
-def kmers_k(primer, k):
-	k = int(k)
-	kmers = []
-	for i in range(len(primer)-k+1):
-		kmer = primer[i:i+k]
-		kmers.append(kmer)
-	return(kmers)
-
-#Returns the hamming distance between two strings
-def hamming(s1, s2):
-    assert len(s1) == len(s2)
-    return sum(ch1 != ch2 for ch1, ch2 in zip(s1, s2))
+from feature_search import *
 
 def merge_reads(r1,  r2, library_name, run_name, bin, data):
